backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-dotenv
psycopg2-binary
pydantic-settings
